source code, required libraries, and executable.
The .py is the source code for the app, while the reuirement.txt contains the libraries that was installed in my virtual environment that I used to code the app. The Zip file contains the Windows executable of the app. After extraction do not move the exe out of the folder.